+++ CHANGELOG.md
@@ -1,8 +1,11 @@
 # Changelog
 
-### 0.0.35:
+### 0.0.35: BREAKING CHANGE
 * Make line comments emit a single newline. This fixes the behavior of
   lines which end in line comments. like `foo // comment\n`
+* Require line comments to be preceded or followed by a whitespace or
+  beginning/end of input. This is needed to allow writing things like
+  URLs which use the `//` sequence.
 
 ### 0.0.34: BREAKING CHANGE
 * Remove built-in markdown support. This has long been an outlier